move getMEssages to locale folder
